FEAT: (#6) 자료형 BigDecimal 로 변경
- Double 의 부동 소수점 문제를 우려하여 BigDecimal 로 변경하였습니다.

+++ src/main/java/com/meetup/server/startpoint/domain/type/Location.java
@@ -5,14 +5,16 @@ import jakarta.persistence.Embeddable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Embeddable
 @NoArgsConstructor(access = lombok.AccessLevel.PROTECTED)
 @Getter
 public class Location {
 
     @Column(name = "road_longitude", length = 255, nullable = false)
-    private Double roadLongitude;   //경도
+    private BigDecimal roadLongitude;   //경도
 
     @Column(name = "road_latitude", length = 255, nullable = false)
-    private Double roadLatitude;    //위도
+    private BigDecimal roadLatitude;    //위도
 }
